requirements.txt
tensorflow==2.15.0
streamlit==1.31.0
pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.0